added mesh optimization (reuses vertices that are adjacent) increases speed of geometry & lighitng

+++ src/gfx/mesh.h
@@ -245,7 +245,7 @@ private:
     vector<GFXVertex> *active_list;
     vector<int> *active_ind;
     vector <Mesh *> lod;
-    vector <int> lodsize;
+    vector <float> lodsize;
     GFXVertex vertex;
     GFXMaterial material;
     int faction;
